removing inactive users pockeyt lite

+++ app/Console/Commands/RemoveUserPockeytLite.php
@@ -45,40 +45,42 @@ class RemoveUserPockeytLite extends Command
   public function handle()
   {
     $timeNow = Carbon::now();
-    $timeLimit = $timeNow->subMinutes(20);
+    $timeLimit =  Carbon::now()->subMinutes(20);
 
-    $userLocations = Location::whereNotBetween('updated_at', [$timeLimit, $timeNow]);
+    $userLocations = Location::whereNotBetween('updated_at', [$timeLimit, $timeNow])->get();
 
-    foreach ($userLocations as $userLocation) {
-    	$business = Profile::findOrFail($userLocation->location_id);
-    	if ($business->account->pockeyt_lite_enabled) {
-    		$squareLocationId = $business->account->square_location_id;
-        $itemId = $business->account->square_item_id;
-        $variationId = 'pockeyt' . $userLocation->user_id;
-        $squareToken = $business->square_token;
+    if (count($userLocations) > 0 ) {
+	    foreach ($userLocations as $userLocation) {
+	      $business = Profile::findOrFail($userLocation->location_id);
+	      if ($business->account->pockeyt_lite_enabled) {
+	        $squareLocationId = $business->account->square_location_id;
+	        $itemId = $business->account->square_item_id;
+	        $variationId = 'pockeyt' . $userLocation->user_id;
+	        $squareToken = $business->square_token;
 
-        try {
-          $token = Crypt::decrypt($squareToken);
-        } catch (DecryptException $e) {
-          dd($e);
-        }
+	        try {
+	          $token = Crypt::decrypt($squareToken);
+	        } catch (DecryptException $e) {
+	          dd($e);
+	        }
 
-        $client = new \GuzzleHttp\Client(['base_uri' => 'https://connect.squareup.com/v1/']);
+	        $client = new \GuzzleHttp\Client(['base_uri' => 'https://connect.squareup.com/v1/']);
 
-        try {
-          $response = $client->request('DELETE', $squareLocationId . '/items' . '/' . $itemId . '/variations' . '/' . $variationId , [
-            'headers' => [
-              'Authorization' => 'Bearer ' . $token,
-              'Accept' => 'application/json'
-            ]
-          ]);
-        } catch (RequestException $e) {
-          if ($e->hasResponse()) {
-            return $e->getResponse();
-          }
-        }
-        $userLocation->delete();
-    	}
-    }
+	        try {
+	          $response = $client->request('DELETE', $squareLocationId . '/items' . '/' . $itemId . '/variations' . '/' . $variationId , [
+	            'headers' => [
+	              'Authorization' => 'Bearer ' . $token,
+	              'Accept' => 'application/json'
+	            ]
+	          ]);
+	        } catch (RequestException $e) {
+	          if ($e->hasResponse()) {
+	            return $e->getResponse();
+	          }
+	        }
+	        $userLocation->delete();
+	      }
+	    }
+	  }
   }
 }
